move neutron vault side logic to separate file
